2023-5-4

+++ model_adni.py
@@ -103,7 +103,7 @@ class AE_adni(nn.Module):
                 optimizer.zero_grad()
 
                 # self-reconstruction loss
-                input_ = Variable(image).to(self.device)
+                input_ = Variable(image).to(self.device).float()
                 reconstructed, z, zu, zv = self.forward(input_)
                 self_reconstruction_loss = self.loss(input_, reconstructed)
